Address PR comments

+++ src/main/java/com/sfjs/data/BaseJobListingData.java
@@ -31,5 +31,5 @@ public class BaseJobListingData extends BaseData {
   @Getter @Setter String location; //?: string;
   @Getter @Setter String country; //?: string;
 //  Boolean jobIsBeingEdited; //?: boolean;
-  @Getter @Setter Long roundNumber;
+  @Getter @Setter Integer roundNumber;
 }
